Add support to run the model on CPUs
This change adds support to run the model using vLLM on CPUs. This
makes it easier to run it on dev machines without GPUs. The latest
version of transformer (v4.44) requires a chat template to be specified for
models which don't have it in the tokenizer. So, that is included in
this change as well.